@@ -1,5 +1,5 @@
 {
-  "name": "goit-react-hw-08-phonebook",
+  "name": "Phonebook",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -11,16 +11,17 @@
     "modern-normalize": "^1.0.0",
     "normalizr": "^3.6.1",
     "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-icons": "^4.1.0",
+    "react-modal": "^3.16.1",
     "react-redux": "^7.2.2",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.0.0",
     "react-scripts": "4.0.1",
     "react-toastify": "^6.2.0",
-    "slugify": "^1.4.6",
     "redux-persist": "^6.0.0",
     "shortid": "^2.2.15",
+    "slugify": "^1.4.6",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
@@ -53,6 +54,7 @@
     "husky": "^4.3.0",
     "lint-staged": "^10.5.2",
     "netlify-cli": "^6.9.2",
-    "prettier": "^2.2.0"
+    "prettier": "^2.2.0",
+    "@babel/plugin-proposal-private-property-in-object": "^7.14.0"
   }
 }
