@@ -1,24 +1,15 @@
 /** @format */
 
 import React from "react";
-import ReactDOM from "react-dom";
-import { BrowserRouter } from "react-router-dom";
-import { Provider } from "react-redux";
-import { PersistGate } from "redux-persist/integration/react";
+import ReactDOM from "react-dom/client";
 import App from "./App";
-import { store, persistor } from "./redux/store";
 import "modern-normalize/modern-normalize.css";
 import "./index.css";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <Provider store={store}>
-      <PersistGate loading={null} persistor={persistor}>
-        <BrowserRouter>
-          <App />
-        </BrowserRouter>
-      </PersistGate>
-    </Provider>
-  </React.StrictMode>,
-  document.getElementById("root")
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
+ <React.StrictMode>
+   <App/>
+  <div>100000000000000000000000</div>
+ </React.StrictMode>
 );
